acow

+++ src/main/java/com/sast/atSast/service/ContestService.java
@@ -9,19 +9,26 @@ import java.util.List;
 public interface ContestService {
 
     void createContest(Contest contest);
+
     void updatepushLink(long contestId, String pushLink);
+
     List<TeamMember> getTeamById(long contestId);
+
     List<Contest> getContestByCurr(Integer curr);
+
     List<Contest> getContest();
+
     Contest getContestById(Integer id);
-    void updateCurr(Integer contestId,Integer curr);
-    void updateComment(Integer contestId,String comment);
-    void updateJudge(Integer judging,long contestId);
+
+    void updateCurr(Long contestId, Integer curr);
+
+    void updateComment(Long contestId, String comment);
+
+    void updateJudge(Integer judging, long contestId);
 
     Contest getContestById(long contestId);
 
     List<Contest> findAll();
-
 
 
 }
