+++ src/main/java/com/sast/atSast/service/JudgesResultService.java
@@ -10,5 +10,5 @@ public interface JudgesResultService {
 
     JudgesResult getResult(long teamId, long contestId, long judgeUid);
 
-    List<JudgesResult> getResults(Integer contestId);
+    List<JudgesResult> getResults(Long contestId);
 }
