+++ src/main/java/com/sast/atSast/service/TeamMemberService.java
@@ -6,7 +6,9 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface TeamMemberService {
-    List<TeamMember> getTeams(Integer contestId);
-    void updateTeam(Long teamId,Integer score,String result);
+    List<TeamMember> getTeams(Long contestId);
+
+    void updateTeam(Long teamId, Integer score, String result);
+
     void insertTeam(TeamMember teamMember);
 }
